+++ Z1R_Tracker/Z1R_WPF/DungeonUI.fs
@@ -207,16 +207,17 @@ let makeDungeonTabs(cm:CustomComboBoxes.CanvasManager, posY, selectDungeonTabEve
         let oldManCountTB = new TextBox(IsHitTestVisible=false, BorderThickness=Thickness(0.), FontSize=12., Margin=Thickness(0.), Width=45.,
                                         HorizontalContentAlignment=HorizontalAlignment.Center, Foreground=Brushes.Orange, Background=Brushes.Black)
         let oldManBorder = new Border(Child=oldManCountTB, Width=45., BorderThickness=Thickness(0.), Background=Brushes.Black)
-        oldManBorder.ToolTip <- "'Old Man Count' - the number of NPC-with-hint and Bomb-Upgrade you have marked,\nand the total number expected in this dungeon"
+        oldManBorder.ToolTip <- "'Old Man Count' - the number of 'old men'\n(NPC-with-hint/Bomb-Upgrade/Hungry-Goriya/\nLife-or-Money) rooms you have marked, and\nthe total number expected in this dungeon"
+        ToolTipService.SetShowDuration(oldManBorder, 8000)
         let updateOldManCountText() = 
             if TrackerModel.IsHiddenDungeonNumbers() then
                 if TrackerModel.GetDungeon(level-1).LabelChar <> '?' then
                     let i = int(TrackerModel.GetDungeon(level-1).LabelChar) - int('1')
-                    oldManCountTB.Text <- sprintf "OM:%d/%d" oldManCount (TrackerModel.GetOldManHintCount(i))
+                    oldManCountTB.Text <- sprintf "OM:%d/%d" oldManCount (TrackerModel.GetOldManCount(i))
                 else
                     oldManCountTB.Text <- sprintf "OM:%d" oldManCount
             else
-                oldManCountTB.Text <- sprintf "OM:%d/%d" oldManCount (TrackerModel.GetOldManHintCount(level-1))
+                oldManCountTB.Text <- sprintf "OM:%d/%d" oldManCount (TrackerModel.GetOldManCount(level-1))
         updateOldManCountText()
         if TrackerModel.IsHiddenDungeonNumbers() then
             TrackerModel.GetDungeon(level-1).HiddenDungeonColorOrLabelChanged.Add(fun _ -> updateOldManCountText())